+++ cmd/ssql/main.go
@@ -68,6 +68,7 @@ func buildRootCommand() *cf.Command {
 	cmd = commands.RegisterTop(cmd)
 	cmd = commands.RegisterDistinct(cmd)
 	cmd = commands.RegisterCount(cmd)
+	cmd = commands.RegisterDescribe(cmd)
 	cmd = commands.RegisterWhere(cmd)
 	cmd = commands.RegisterUpdate(cmd)
 	cmd = commands.RegisterCast(cmd)